fix: clang and delete mains
