--- memchr.c
@@ -1,21 +1,0 @@
-#include <stdio.h>
-#include "s21_string.h"
-
-void *s21_memchr(const void *arr, int ch, s21_size_t n) {
-    void *result = S21_NULL;
-    while (n > 0) {
-        const unsigned char * arr_ch = (const unsigned char *)arr;
-        if (*arr_ch == ch) {
-            result = (void *)arr_ch;
-            break;
-        }
-        n--;
-    }
-    return result;
-}
-
-int main() {
-    char * arr = "abc";
-    void * pt = s21_memchr(arr, 'g', 3);
-    printf("%p", pt);
-}
